Starting moving to rustc stable + actix
Note the the rusqlite dependency had to be patched with an override, as it does not compile on stable just yet. Need to investigate this after the migration is complete.

+++ src/config/config.rs
@@ -158,7 +158,7 @@ pub struct ConfigHTTP {
     pub inet: SocketAddr,
 
     #[serde(default = "defaults::http_workers")]
-    pub workers: u16,
+    pub workers: usize,
 
     pub record_token: String,
 }
